requirements.txt
# Insurance Broker AI — Dependencies
chainlit>=2.10.0
google-genai>=1.0.0
python-dotenv>=1.0.0

# MCP Server
fastmcp>=2.14.5
mcp>=1.25.0
pydantic>=2.12.0
pydantic-settings>=2.0.0
jinja2>=3.1.0
aiosqlite>=0.21.0

# Admin panel (FastAPI)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Export: PDF / XLSX / DOCX
openpyxl>=3.1.0
python-docx>=1.0.0
weasyprint>=62.0
